Required changed from #23
Required changes from #23

Note: I did try to follow the guidelines, it just wouldn't build.

Atomic changes are great and can follow guidelines for atomic pull requests in the future. This is a demon from the closet of the past pull requests, and needs to be resolved.

Kindly let Sergei reject the changes that concern him. I am sure he is capable of discussion. If no pull requests are approved, the project becomes difficult to coordinate. We have no other development branch to merge to and coordinate, only the main branch.